third_party/WebKit/LayoutTests/http/tests/devtools/console/console-focus-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that interacting with the console gives appropriate focus.

Message count: 2

Running: testClickingWithSelectedTextShouldNotFocusPrompt
Focused element: BODY
Viewport scrolled to top: true
Clicking message 0
Focused element: BODY
Viewport scrolled to top: true

Running: testClickOnMessageShouldFocusPromptWithoutScrolling
Focused element: BODY
Viewport scrolled to top: true
Clicking message 0
Focused element: TEXTAREA
Viewport scrolled to top: true

Running: testClickOutsideMessageListShouldFocusPromptAndMoveCaretToEnd
Focused element: BODY
Viewport scrolled to top: true
Selection before: {"startLine":0,"startColumn":1,"endLine":0,"endColumn":1}
Clicking on container
Focused element: TEXTAREA
Viewport scrolled to top: false
Selection after: {"startLine":0,"startColumn":6,"endLine":0,"endColumn":6}
